+++ changelog.md
@@ -1,2 +1,11 @@
+# v1.0.1
+
+Fixed permission issues on windows
+
+Added windows install script
+
+
+
 # v1.0.0
+
 Working prototype. Builds for win64 and arm64 are working.